requirements.txt
discord~=2.3.2
python-dotenv~=1.0.1
numpy~=1.26.4